Added code to optionally include Shipping Charges in the cost. Also added new field to orders table 'shipping_cost'
All of the new code can be disabled by define SHIPPING_FREE 1

All places where shipping charges may be displayed it will always
display if a charge was entered when the order was placed.
(This is case orders were made when there were shipping charges and
then the store was changed to free shipping)

mysql_catalog.sql was updated to include extra records for admin package
as well as shipping_cost (all changes were made to orders table)

To upgrade you may be able to DROP TABLES orders and then just execute
the commands from mysql_catalog.sql that create the orders table.