Bug Fix: When updating a product there would be product_to_manufacturer and products_to_categories that would have the product number cleared to zero.
The problem is that there was a check for inserting new records for each
that would be executed each time a product was updated (even if the manfact/category
lists were un-changed).


Harald check this one over closely, I tested it quite a bit so am pretty sure it is ok.

tom