These changes will allow the following fields to be optional in the customer information and address book entries:
Gender, Date of Birth, Suburb and/or State

They are enabled (or disabled) in application_top.php via defines for:

ACCOUNT_GENDER, ACCOUNT_DOB, ACCOUNT_SUBURB and AACCOUNT_STATE

The values used must be 0 for off and 1 for On (used in math for formatting)

The address book interface from admin has not been updated yet.
I'll do that tonight.

There were a lot of changes here I wanted to get in before I got a conflict..

Tom