Add Prev/Next navigation bar
Adding a Prev/Next navigation bar to search.php using a set
of 3 generic functions that can be used in most scripts that
potentially display a large list of rows such as the products.php
in the admin package.  This version of the navigation bar differs
from the previous implementation in the following areas:

(1) instead of executing the query and then use tep_db_num_rows to
retrieve the row count; it would reconstruct the query to use
"select count(*)" to determine the affected row count

(2) it uses section variable to store the row count and hence reduces
unnecessary re-counting in subsequent pages

(3) it allows users to determine the max number of page links to
be displayed; when the number of pages exceeds this maximum, previous
and/or next "window" buttons would be displayed to allow users to
scroll to the previous/next set of page links (this feature is necessary
to support large number products)

(4) instead of display number of items found, it displays the range
being displayed as well. e.g. Products 16 to 30 of 561 found

Please note that the MAX_DISPLAY_SEARCH_RESULTS has been intensionally
changed to a small number to demonstrate the prev/next window feature.
Please adjust MAX_DISPLAY_SEARCH_RESULTS and MAX_DISPLAY_PAGE_LINKS
to suite your database.

Please also note that my German translation of the text used might not be
100% accurate.