Added new module: also_purchased_products.php
A new 'Customer who bought this product also purchased' list
is added to bottom of the product info page.

MIN_DISPLAY_ALSO_PURCHASED and MAX_DISPLAY_ALSO_PURCHASED are
used to control the list.

MIN_DISPLAY_ALSO_PURCHASED determines the minimum number of
purchased items used to randomly generate the final list of
products displayed.  This constant is initually set to 1 for
testing purpose and should be set to higher number for better
randomizaton.  If not enough products are ordered, the list
would not be displayed.

MAX_DISPLAY_ALSO_PURCHASED determines the maximum number of products
to display.