Initial Entry into CVS.
A tool to upgrade a TEP database to use
one way MD5 encrypted passwords.
It will change the size of the customers_password
field in the customers table to be
a varchar(4) and encrypte the users old
password and update it in the database.
A simple check is included to try to prevent it from being
run more than once on a database.