Added modular payments. I havent really written much code for this, but just moved things around. All Payment modules are kept in a folder (by default catalog/includes/modules/payments) and the programming has moved from the checkout files to each module.
Make sure you get the folder right for the modules, or nothing will work, specially from
the admin tool. There is a new define called PAYMENT_MODULES that is automatically updated
from the module page of admin tool.


From the admin tool you can install/remove modules, and this will add/remove the
necesary defines from the database. Once a module is installed you can go to the
payment options section to configure each payment.

Language define still need to be cleaned up. At the moment they are at the top of each
module.


Dont forget that this is CVS, and that means testing area. Do not try this on your
production store until it is debugged.

... and send me or the list any bugs you find and ill fix them.

I have created sample module (located in catalog/includes/modules/payment/sample.php)
that is properly commented. This is the place to start if you want create a new module
or have an idea of how they work.