Closed some <option> tags. & fixed a type in english and a minor javascript fix to the state text field.
it now works on netscape 3 aswell. Let me know if you have problems with this javascript. But please, send me a detailed description of the problem and tell me what browser you are using. ... and what version of the browser aswell.

Theres been some complaints about javscript in the mailing list, but withtout this information its impossible to debug.