Added Order Comments support this DOES require an extra field on the orders table.
The customer will b able to enter comments when the order is placed and then
can view the comments in order_history_info. The admin can edit the comments
from orders.php in admin. The user will be able to follow the order progress.

Many thanks to Haim Barad for the initial code.