Next round of updates for modified checkout process.
checkout_payment now has DELIVERY Address and a CHANGE ADDRESS Link.

Yes it is just a Text Link, not a button.

If someone (PLEASE) adds a button please leave the text as a descrption
of the btton. I do not have much access to graphics, etc.

The checkout link on the left of tool bar takes you to checkout_address
and the checkout on the right of the tool bar takes you to checkout_payment