Changed the order of parameters in the tep_image function.
Mainly, to easily ommit width and height information.

Also, the width and height has been removed from all calls to tep_image
so that changing the buttons wont need hacking the code to get the sizes right.

2 more configs have moved in the database:

IMAGE_REQUIRED - Enable this if you want to see a broken image when an image doesnt exit. If you disable and the image doesnt exist, nothing is returned. This used to be in application_top

CONFIG_CALCULATE_IMAGE_SIZE - With this setting on, image sizes are calculated if set to zero (height, width or both). For an example, try setting SMALL_IMAGE_WIDTH to 200 and set SMALL_IMAGE_HEIGHT to 0. Make sure CONFIG_CACULATE_IMAGE_SIZE is set to 1 and you should get all product images with the width set to 200 and the height calculated proportionally so that aspect ratio is not lost.