IMPORTANT:
SHIPPING_MODULES is now a list of filenames separated by semicolons, not spaces.

So anyone that is running more than one shipping module in their store
will have update this variable from the admin tool.

If you click on "Shipping" in the "Modules" section in the admin tool, it is
automaticaly updated for you.


I did this so that both payment & shipping modules use the same syntax.
I think a semi-colon is better, because filenames can have spaces.