These are the final changes to the authorize net payment module. I think it works properly now.
NOTE: You need to set HTTP_SERVER to use this module. You cant leave it empty.
This also applies to HTTPS_SERVER if ENABLE_SSL is set. Which is a "must" in production stores.

PS: It could do with a bit more work displaying messages and storing confirmation values, but
theres no framework for it at the moment. However, its a start and it works with the 'testing' account (default).