Increased timeout in backup procedure to 3 minutes as suggested by Reini Urban.
I made a tep_set_time_limit for this because you cant use this function when safe mode is activated.

Fixed a bug with the number of fields by turning $table_list to an array.
Also fixed the default value when it is '0'.
And removed the display of page parse time in the backup file.