Added the Search Engine Friendly URL's functionality.
Please report any bugs that you find.

You can enable it in the configure.php file.

The only URL's that arent slashed are a result of GET forms (manufacturers box and currencies box). We will address that problem shortly.

Also, we need to include a <base href=... to make sure the images are loaded correctly. (coming next)