We've had an experience with a spider in catalink, it kept spawning new sessions and the spider thought they were different paths. It stayed there for hours!
Adding the SID as a regular variable seems to work ok, so if you had cookies disabled, your URL's would you like this:
http://exchange/catalog/default.php/products_id/8?sID=xxxxxxxxxxxxxxxxxx

This seems to solve the problem, but we are still waiting to see what the spider does next time he comes!

Ive added an extra parameter to be able to generate links without SID. This can be useful for email_orders, or get forms where SID has to be included anyway as a hidden field (manufacturers box).

Ive added some comments so we dont get lost, this is getting tricky.