Ive split CURRENCY_VALUE in 2: CURRENCY_VALUE and DEFAULT_CURRENCY
CURRENCY_VALUE is the base currency used to store data in the database.
and DEFAULT_CURRENCY is used for display monetary numbers.

Before, CURRENCY_VALUE was being used for both.

Ive made CURRENCY_VALUE the default for the base currency when calculating exchange rates
And DEFAULT_CURRENCY is the default for tep_currency_format().

I still have to look at how products are added/edited.
For now, I would use the same value in both of these.