Now that the new url scheme has been tested for a while we can see that there are cases when it cant be used.
One example of this is the preconfirmationcheck; when it fails, it redirects
the user to checkout_payment again with an error message in the URL. This
long error message (with apostrophes, quotes, etc) in the URL does not work
properly when using Search Engine Friendly Urls.

Another example is in product_info, if the URL for the product contains
slashes, it doesnt work either.

If you look at both examples, you can see that neither of the cases really
need the friendly urls there. Friendly URLs are mainly targeted to search
engines and a search engine should never get to preconfirmation check. And
if he doesnt follow the URL for a product we dont really care either.

A possible solution could be to add an extra parameter to tep_href__link().
This extra parameter would control wether or not to transalate the urls to
friendly mode, and its value would default to the setting in
application_top.php.