* enhanced the UI for the administration tool - now simple true/false parameters can be controlled via radio buttons in the info box - neat!
i first tried with checkboxes, but a value needs to be sent even if the
parameter is false - radio buttions is perfect for this. this should be
further enhanced to allow other true/false combinations, for example
ascending and descending.

* moved the USE_DEFAULT_LANGUAGE_CURRENCY parameter into the database

NOTE: at the start we used "if (constant == 'true')" and then moved to
"if (constant)" - we should move back to the first method to ensure full
compatibility with the php versions, and to ensure the "if" statement
does what it is meant to do. this can be discussed further if needed. i
will post this to the developers mailing list tomorrow, and also update
the STANDARDS file.