the plain old backup section has now been turned into a full featured backup manager (ok, not yet - restoration comes soon =))
there is a new directory called DIR_FS_BACKUP which is located in
admin/backups; this is where the backup files will be placed. apache needs
write permission in this directory. there is a file in this directory
called 'emptyfile' and it is just that - this is just for the cvs to create
the directory in your workspace.