There are several variables carried around the checkout process with hidden fields in forms. These have to be passed on the payment server too, along with the session id.
These changes move the $sendto variable from the forms to the session. Theres no need to carry it around anymore (simplifying payment modules) and its value will stick until checkout is processed (we un-register the variable then).

I'll be moving the rest of the variables so that payment modules will only have to carry the session id.