tep standards updates (more to come)
added a new function -> tep_get_category_tree() which puts the category
list in tree view in an array

this new function now works fine when moving a category

NOTE: the possibility of editing a category's parent ID has been removed,
which is taken care of from the (category) Move feature. It is now possible
to move categories to the top level.