Right now, the boxes class puts the form tags either between <table> and <tr> or between <tr> and <td>
These places are illegal (there is also a bug report about this).

Ive changed it so that form tags are displayed either outside the <table> tag or inside the <td>

The reason this was done this way was to eliminate some whitespace that form tags produces, this will solved in the stylesheet now, next commit.