This is related to bug #71
When an empty var is added to the URL, and Search Engine URL is enabled, double slashes ended up messing the variables.
Ive fixed it tep_get_all_get_params() by checking if the variable has a value before adding it to the URL.


There was also a bug checking date because spanish date format has 'aaaa' to represent the year.
Change tep_check_date() to accept spanish year notation too, maybe this should be looked into with german too.