introducing a new logger class, that writes messages to a log file
right now the page parse time module is taking advantage of it -> because
of this EXIT_AFTER_REDIRECT and tep_exit() have been removed.

tep_redirect() has some new logic to determine whether or not to write the
page parse time before calling the exit() call.

the write() method can be extended to write to a database table.