Initial commit of HTML admin invoice and packing slip.
They simply remove header and footer and column left/right and some other
minor formatting differences. otherwise the content is the same as
the order details.

The language files need to be translated and the buttons aren't quite the right
color. There is also a new configuration variable which gives the company
name and address for the header og the form.