First set of changes to get rid of the rounding tax problem.
The tep_get_tax_rate() will now take care of deciding which country, zone to use for tax displaying.
Ive changed the order of the parameters so they can be ommited and assigned a default value.

Ive added a couple of helper functions,
one to calculate tax rounded to as many decimals as the currency has.
and another one that adds the tax or not based on the DISPLAY_PRICE_WITH_TAX parameter.

Next I will commit the changes to the rest of the files, I havent been able to test this under all circumstances.

Feel free to open a bug report if you see anything wrong.