Update to standards (including apostrophe support) in create account and account edit.
Introducing a new way of handling zones, still needs some polishing and the address book isnt done yet.

The basic idea is to let the user type in the state/province (removing all the javascript with the zones so that we can finally add as many zones as we like to the distribution) in the next page some checking is done and one of 3 things can happen:

1) the country has no zones, so we just keep the text typed in.
2) the country has zones and the user typed in an exact match, we save the zone_id
3) the country has zones and we dont have an exact match, we display a dropdown with zones for the country to choose from.

Its a first draft and can/will be improved on, any suggestions welcome.