the new logic for windows server handling short dates prior 1970 has been updated to handle all dates
thanks to Michael Burke for forwarding the fix to David Garcia Watkins

here is what Michael had to say:

The problem is that the valid range of a timestamp in PHP terms is
typically from Fri, 13 Dec 1901 20:45:54 GMT to Tue, 19 Jan 2038 03:14:07
GMT. (These are the dates that correspond to the minimum and maximum values
for a 32-bit signed integer). On WIN32 systems this range is limited from
01-01-1970 to 19-01-2038.

So, the current code does not work if the DOB is after January 19th of any
year earlier than 1970. In those cases an error condition will be raised.
To cover all month/day combinations, it would be better to use 2037 rather
than 2038 - the last year where all month/days are covered. The key being
the top of the timestamp range being only the 19th of January. The example
is June 18, 1951. The routine as it currently stands would result in a date
function call of "date(DATE_FORMAT, mktime($hour, $minute, $second, 6, 18,
2038)));" This is an invalid PHP date being greater than 19 Jan 2038
03:14:07 GMT.

Using 2037 rather than 2038 results in the routine working for all dates
(perhaps leap years might be a problem, but that would be a problem
regardless).