added a security layer to the user-input parameters of the tep_draw_*_field functions.
Using htmlspecialchars() would be ideal to use, but is an overkill in most cases. We found the only character needing translating was the double quote character, it is important this character is translated so that the parameters that contain user-input do not break the html tag. For example:

Invalid:
<input type="text" name="name" value="boo"hoo">

Valid:
<input type="text" name="name" value="boo&quot;hoo">

A new tep_parse_input_field_data() has been introduced to take care of "protecting" the user-input parameters, which currently does so with the use of strtr().

With strtr() we are able to define which characters should be translated. All but currently one call to this function only translate the doube quotes character. The call in the tep_draw_pull_down_menu() function translates not only the double quote character, but also single quotes, less than, and greater than characters.

This is basically using htmlspecialchars() but the ampersand character (&) is not being translated, as data passed to the tep_draw_pull_down_menu() function may have gone through htmlspecialchars() already - thus producing an effect of &amp;nbsp; which has existed in the advance search page for a while (the categories pull down menu).

Please provide feedback if problems are experienced with this update.