Fixed the copyright announcement to be in compliance with section 2c of the GNU General Public License.
A proposal was made a while ago on adding such a compliant copyright
announcement to the project, and is being realized in best interest of the
project.

The proposal was discussed on the online forums, and can be read here:

http://forums.oscommerce.com/viewtopic.php?t=24580

Having such a compliant copyright announcement means that it cannot be
removed, or only appropriately modified.

This is only for the Administration Tool of osCommerce. The Catalog module
is not affected by this proposal.

A Frequently Asked Questions entry has been created for this issue, and can
be read here:

http://www.oscommerce.com/community.php/faq,26/q,50