Quote from http://www.searchengineworld.com/spiders/google_faq.htm#item355
"Does Google index dynamic content?
A: It will in certain instances. What criteria is used to determine if a dynamic page is indexable is debatable. Most have found that
clean, high ranking (High page rank) sites can get dynamic content indexed"

This makes me think that Googlebot does indeed index dynamic pages, but not everybodies.

Another Quote:

"Sites that use session tracking urls to give each visitor a dynamic url. These sites can generate an infinite amount of pages for a
spider to visit. These types of pages, are usually blocked from being indexed by Google."

With this modification, the session id is back to the parameter area (even with Search Engine Friendly URL's turned on) so that no
spiders get trapped in our infinite number of pages and we are not banner from Google.

Any extra info on the subject is welcome.