Fix bug report 806
Now that the order class is being instantiated before the payment method
before_process() method is being called, the new credit card number has
to directly overwrite the value in the order class instance if filtering
the middle digits of the credit card number is enabled, as the order class
instance uses the values from HTTP_POST_VARS - which this module was
updating (and thus no longer being stored in the order class instance as it
is now being instantiated before before_process()).

Confusing? Alternate response: filtering of the middle credit card number
digits is now working again.