add a new update_status() method to the payment modules
this solves a chicken and egg problem with the order class and payment
class where both need information from one another when instantiated in
the checkout_confirmation.php page

the order class needs the payment module title and the payment module needs
the order destination information for the new dynamic status feature when
the payment module is assigned a zone

the solution to the chicken and egg problem is to instantiate the payment
class first then the order class afterwards

a call to the update_status() method is made manually to set the status of
the payment module dynamically if it has been assigned a zone. the logic
used in the update_status() method was otherwise executed in the class
constructor when instantiated (which is still being done on the
checkout_payment.php page where the chicken and egg problem does not exist
as the order class does not utilize the payment module title here; this is
done with the constructor calling the update_status() method automatically)

the new update_status() method does not break the modules in the
contributions section as the main payment class uses logic to detect if the
method exists or not - if the method does not exist then the module has
obviously not been updated to take advantage of the new dynamic status
feature. this may change when 2.2 goes gold.