Commit workboard 39 - Order commends work.
It puts the order comments box back to before Process. In fact
it is on both checkout_shipping and checkout_payment.

We may re-look at the designs of these pages at a later date.

Also moves the order comments from the order table to
the order_status_history table so that comments can be
made with status update.