I found Linda's bug.
The comment on line134 says we're only checking one attribute, but the index used is $i
which will be 1 the second time through. So if the second product only has one attribute
then we will do a query and get '' returned and '' -1 = -1 so we are out of stock

Fixed the index to be 0 all the time and also made sure we did get database
rows returned before using the result. (tep_db_num_rows())