The realization of the My Account proposal
The sections of the account area are now separated into:

Order Overview (shown is orders have been made)
Personal Information (account, address book, password)
Orders
Notifications (newsletters, product notifications)

The main account page will become modular post 2.2.

The javascript error messages have been removed - the normal field error
messages are now used instead. We are looking into a solution to easily
administrate which fields should be displayed in a form.

The separation of the account area has introduced new pages - these will
more than likely disappear after the template structure has been
implemented. The new pages have been added to complete the realization
of the proposal and to stay on track of the 2.2 milestone release path.

The limit of address book entries is now including the default address -
previously if the maximum size of 5, a total of 6 entries would exist in
the address_book table. Now it would be 5 (following the example) as it is
now possible for the customer to change the default address.

Code clean-up will be done in the coming days.

(sorry for the big commit, and have fun! ;))